// src/constants/enums.ts
export enum EProficiencyLevel
{
    Beginner = "Beginner",
    Intermediate = "Intermediate",
    Advanced = "Advanced",
    Fluent = "Fluent"
}

export enum EEducationType
{
    BasicEducation = "BasicEducation",
    TechnicalCourse = "TechnicalCourse",
    Graduation = "Graduation",
    PostGraduation = "PostGraduation"
}

export enum EApplicationStatus
{
    ANALISE = "Analysis",
    APROVADO = "Approved",
    REPROVADO = "Disapproved",
    DINAMICA = "Activity",
    ENTREVISTA = "Interview"
}

export enum EEducationStatus
{
    Incomplete = "Incomplete",
    Underway = "Underway",
    Complete = "Complete"
}